@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
@@ -9,11 +9,22 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="7.0.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.1" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
-    <PackageReference Include="Neoron.API.Validation" Version="1.0.0" />
-    <PackageReference Include="Neoron.API.DTOs" Version="1.0.0" />
-    <PackageReference Include="Neoron.API.Interfaces" Version="1.0.0" />
-    <PackageReference Include="Neoron.API.Repositories" Version="1.0.0" />
+    <PackageReference Include="Serilog.AspNetCore" Version="6.0.0" />
+    <PackageReference Include="Serilog.Exceptions" Version="8.0.0" />
+    <PackageReference Include="Serilog.Enrichers.Environment" Version="2.2.0" />
+    <PackageReference Include="Serilog.Enrichers.Process" Version="2.0.2" />
+    <PackageReference Include="Serilog.Enrichers.Thread" Version="3.1.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Neoron.API.Validation\Neoron.API.Validation.csproj" />
+    <ProjectReference Include="..\Neoron.API.DTOs\Neoron.API.DTOs.csproj" />
+    <ProjectReference Include="..\Neoron.API.Interfaces\Neoron.API.Interfaces.csproj" />
+    <ProjectReference Include="..\Neoron.API.Repositories\Neoron.API.Repositories.csproj" />
   </ItemGroup>
 
 </Project>
