@@ -1,119 +1,94 @@
-using System.Collections.Concurrent;
+using System;
+using System.Diagnostics;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Neoron.API.Extensions;
+using Neoron.API.Models;
 
 namespace Neoron.API.Middleware
 {
-    public class RateLimitingOptions
-    {
-        public int MaxTokens { get; set; } = 100;
-        public int RefillRate { get; set; } = 10;
-        public TimeSpan CleanupInterval { get; set; } = TimeSpan.FromHours(1);
-    }
-
+    /// <summary>
+    /// Middleware for rate limiting requests.
+    /// </summary>
     public class RateLimitingMiddleware : IDisposable
     {
-        private readonly RequestDelegate _next;
-        private readonly ILogger<RateLimitingMiddleware> _logger;
-        private readonly RateLimitingOptions _options;
-        private readonly Timer _cleanupTimer;
-        private static readonly ConcurrentDictionary<string, TokenBucket> _buckets = new();
+        private static readonly TimeSpan CleanupInterval = TimeSpan.FromHours(1);
 
-        public RateLimitingMiddleware(
-            RequestDelegate next,
-            ILogger<RateLimitingMiddleware> logger,
-            IOptions<RateLimitingOptions> options)
+        private readonly RequestDelegate next;
+        private readonly ILogger<RateLimitingMiddleware> logger;
+        private readonly RateLimitingOptions options;
+        private readonly TokenBucket tokenBucket;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="RateLimitingMiddleware"/> class.
+        /// </summary>
+        /// <param name="next">The next middleware in the pipeline.</param>
+        /// <param name="logger">The logger instance.</param>
+        /// <param name="options">The rate limiting options.</param>
+        public RateLimitingMiddleware(RequestDelegate next, ILogger<RateLimitingMiddleware> logger, IOptions<RateLimitingOptions> options)
         {
-            _next = next ?? throw new ArgumentNullException(nameof(next));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
-            
-            _cleanupTimer = new Timer(CleanupBuckets, null, _options.CleanupInterval, _options.CleanupInterval);
+            ArgumentNullException.ThrowIfNull(next);
+            ArgumentNullException.ThrowIfNull(logger);
+            ArgumentNullException.ThrowIfNull(options);
+
+            this.next = next;
+            this.logger = logger;
+            this.options = options.Value;
+            this.tokenBucket = new TokenBucket(this.options.MaxTokens, this.options.TokenRefillRate);
+
+            // Start the cleanup timer
+            var cleanupTimer = new Timer(Cleanup, null, TimeSpan.Zero, CleanupInterval);
         }
 
-        private void CleanupBuckets(object? state)
-        {
-            var now = DateTime.UtcNow;
-            var staleKeys = _buckets
-                .Where(kvp => (now - kvp.Value.LastUsed).TotalHours > 1)
-                .Select(kvp => kvp.Key)
-                .ToList();
-
-            foreach (var key in staleKeys)
-            {
-                _buckets.TryRemove(key, out _);
-            }
-        }
-
-        public void Dispose() => _cleanupTimer?.Dispose();
-
+        /// <summary>
+        /// Invokes the middleware.
+        /// </summary>
+        /// <param name="context">The HTTP context.</param>
+        /// <returns>A task representing the middleware invocation.</returns>
         public async Task InvokeAsync(HttpContext context)
         {
-            if (context == null)
-            {
-                throw new ArgumentNullException(nameof(context));
-            }
+            ArgumentNullException.ThrowIfNull(context);
 
-            var key = GetClientKey(context);
-            var bucket = _buckets.GetOrAdd(key, _ => new TokenBucket(_options.MaxTokens, _options.RefillRate));
-
-            if (!bucket.TryTake())
+            if (!this.tokenBucket.ConsumeToken())
             {
-                _logger.LogWarning("Rate limit exceeded for client {ClientKey}", key);
                 context.Response.StatusCode = StatusCodes.Status429TooManyRequests;
-                await context.Response.WriteAsJsonAsync(new { error = "Too many requests" }).ConfigureAwait(false);
+                await context.Response.WriteAsync("Too Many Requests").ConfigureAwait(false);
+                LoggingMessages.FailedToAddMessage(this.logger, context.Connection.RemoteIpAddress?.ToString(), new InvalidOperationException("Rate limit exceeded"));
                 return;
             }
 
-            await _next(context).ConfigureAwait(false);
+            await this.next(context).ConfigureAwait(false);
         }
 
-        private static string GetClientKey(HttpContext context)
+        /// <summary>
+        /// Releases unmanaged and - optionally - managed resources.
+        /// </summary>
+        public void Dispose()
         {
-            var clientIp = context.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-            var apiKey = context.Request.Headers["X-API-Key"].ToString();
-            var path = context.Request.Path.ToString();
-            
-            // Include API endpoint in the key to allow different rates for different endpoints
-            return string.IsNullOrEmpty(apiKey) 
-                ? $"{clientIp}:{path}"
-                : $"{clientIp}:{apiKey}:{path}";
+            Dispose(true);
+            GC.SuppressFinalize(this);
         }
-    }
 
-    public class TokenBucket(int maxTokens = 100, int refillRate = 10)
-    {
-        private readonly int _maxTokens = maxTokens;
-        private readonly int _refillRate = refillRate;
-        private double _tokens = maxTokens;
-        private DateTime _lastRefill = DateTime.UtcNow;
-        private readonly object _lock = new();
-        public DateTime LastUsed { get; private set; } = DateTime.UtcNow;
-
-        public bool TryTake()
+        /// <summary>
+        /// Cleans up the middleware state.
+        /// </summary>
+        /// <param name="state">The state object.</param>
+        protected void Cleanup(object? state)
         {
-            lock (_lock)
+            // Perform any necessary cleanup
+        }
+
+        /// <summary>
+        /// Releases unmanaged and - optionally - managed resources.
+        * </summary>
+        /// <param name="disposing"><c>true</c> to release both managed and unmanaged resources; <c>false</c> to release only unmanaged resources.</param>
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
             {
-                RefillTokens();
-                LastUsed = DateTime.UtcNow;
-
-                if (_tokens < 1)
-                {
-                    return false;
-                }
-
-                _tokens--;
-                return true;
+                this.tokenBucket.Dispose();
             }
-        }
-
-        private void RefillTokens()
-        {
-            var now = DateTime.UtcNow;
-            var elapsed = (now - _lastRefill).TotalSeconds;
-            var tokensToAdd = elapsed * _refillRate;
-
-            _tokens = Math.Min(_maxTokens, _tokens + tokensToAdd);
-            _lastRefill = now;
         }
     }
 }
