@@ -3,44 +3,67 @@ using Neoron.API.Models;
 
 namespace Neoron.API.DTOs
 {
-    public record CreateMessageRequest
+    /// <summary>
+    /// Represents a request to create a new Discord message.
+    /// </summary>
+    public class CreateMessageRequest
     {
+        /// <summary>
+        /// Gets or sets the channel identifier where the message will be sent.
+        /// </summary>
         [Required]
-        public required long MessageId { get; init; }
+        public long ChannelId { get; set; }
 
+        /// <summary>
+        /// Gets or sets the guild identifier where the message belongs.
+        /// </summary>
         [Required]
-        public required long ChannelId { get; init; }
+        public long GuildId { get; set; }
 
+        /// <summary>
+        /// Gets or sets the author identifier of the message.
+        /// </summary>
         [Required]
-        public required long GuildId { get; init; }
+        public long AuthorId { get; set; }
 
+        /// <summary>
+        /// Gets or sets the content of the message.
+        /// </summary>
         [Required]
-        public required long AuthorId { get; init; }
+        [StringLength(2000, MinimumLength = 1)]
+        public required string Content { get; set; }
 
-        [Required]
-        [StringLength(2000)]
-        public required string Content { get; init; }
+        /// <summary>
+        /// Gets or sets optional embedded content of the message.
+        /// </summary>
+        public string? EmbeddedContent { get; set; }
 
-        public string? EmbeddedContent { get; init; }
+        /// <summary>
+        /// Gets or sets the type of message.
+        /// </summary>
+        public MessageType MessageType { get; set; }
 
-        public MessageType MessageType { get; init; }
+        /// <summary>
+        /// Gets or sets the creation timestamp of the message.
+        /// </summary>
+        public required DateTimeOffset CreatedAt { get; set; }
 
-        public long? ReplyToMessageId { get; init; }
-
-        public long? ThreadId { get; init; }
-
-        public DiscordMessage ToEntity() => new()
+        /// <summary>
+        /// Converts the CreateMessageRequest to a DiscordMessage entity.
+        /// </summary>
+        /// <returns>A new DiscordMessage entity.</returns>
+        public DiscordMessage ToEntity()
         {
-            MessageId = MessageId,
-            ChannelId = ChannelId,
-            GuildId = GuildId,
-            AuthorId = AuthorId,
-            Content = Content,
-            EmbeddedContent = EmbeddedContent,
-            MessageType = MessageType,
-            CreatedAt = DateTimeOffset.UtcNow,
-            ReplyToMessageId = ReplyToMessageId,
-            ThreadId = ThreadId
-        };
+            return new DiscordMessage
+            {
+                ChannelId = ChannelId,
+                GuildId = GuildId,
+                AuthorId = AuthorId,
+                Content = Content,
+                EmbeddedContent = EmbeddedContent,
+                MessageType = MessageType,
+                CreatedAt = CreatedAt,
+            };
+        }
     }
 }
