@@ -1,42 +1,24 @@
+using System.Globalization;
+using Serilog;
+using Serilog.Exceptions;
+
 namespace Neoron.API.Extensions
 {
     public static class LoggingExtensions
     {
-        public static IHostBuilder AddCustomLogging(this IHostBuilder builder)
+        public static IHostBuilder AddCustomLogging(this IHostBuilder hostBuilder)
         {
-            return builder.UseSerilog((context, services, configuration) => configuration
-                .ReadFrom.Configuration(context.Configuration)
-                .MinimumLevel.Information()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                .MinimumLevel.Override("System", LogEventLevel.Warning)
-                .MinimumLevel.Override("Microsoft.AspNetCore.Mvc.Infrastructure", LogEventLevel.Warning)
-                .Enrich.FromLogContext()
-                .Enrich.WithExceptionDetails()
-                .Enrich.WithEnvironmentName()
-                .Enrich.WithMachineName()
-                .Enrich.WithProcessId()
-                .Enrich.WithThreadId()
-                .Filter.ByExcluding(Matching.FromSource("Microsoft.AspNetCore.StaticFiles"))
-                .WriteTo.Async(a => a.Console(
-                    outputTemplate: "[{Timestamp:HH:mm:ss} {Level:u3}] {Message:lj} {Properties:j}{NewLine}{Exception}"))
-                .WriteTo.Async(a => a.File(
-                    path: "logs/neoron-.log",
-                    rollingInterval: RollingInterval.Day,
-                    retainedFileCountLimit: 7,
-                    outputTemplate: "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] {Message:lj} {Properties:j}{NewLine}{Exception}"))
-                .WriteTo.Async(a => a.Debug(
-                    outputTemplate: "[{Timestamp:HH:mm:ss} {Level:u3}] {Message:lj} {Properties:j}{NewLine}{Exception}",
-                    restrictedToMinimumLevel: LogEventLevel.Debug)));
-        }
-
-        public static Serilog.ILogger CreateBootstrapLogger()
-        {
-            return new LoggerConfiguration()
-                .MinimumLevel.Information()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                .Enrich.FromLogContext()
-                .WriteTo.Console()
-                .CreateBootstrapLogger();
+            return hostBuilder
+                .UseSerilog((context, configuration) =>
+                {
+                    configuration
+                        .ReadFrom.Configuration(context.Configuration)
+                        .Enrich.FromLogContext()
+                        .Enrich.WithMachineName()
+                        .Enrich.WithProperty("Environment", context.HostingEnvironment.EnvironmentName)
+                        .WriteTo.Console(outputTemplate: "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] {Message:lj}{NewLine}{Exception}")
+                        .WriteTo.File("logs/myapp.txt", rollingInterval: RollingInterval.Day);
+                });
         }
     }
 }
